tests: move `TerminalImportSessionSetup` from `tests.test_ui_importer` to `test.helper`
This class is imported by some other test modules.
Thus, it should reside in a module, which is obviously used by other
tests.
